Simplify the `Prism` class
Move to separate modules:
- `highlightAll()`
- `highlightElement()`
- `highlight()`
- `stringify()`
